+++ src/main/java/de/samply/template/graph/factory/ExcelContainerAttributeKeys.java
@@ -1,0 +1,17 @@
+package de.samply.template.graph.factory;
+
+import de.samply.template.AttributeTemplate;
+import de.samply.template.ContainerTemplate;
+import de.samply.template.graph.ContainerAttributeKeys;
+
+public class ExcelContainerAttributeKeys implements ContainerAttributeKeys {
+    @Override
+    public String getContainerKey(ContainerTemplate containerTemplate) {
+        return containerTemplate.getExcelSheet();
+    }
+
+    @Override
+    public String getAttributeKey(AttributeTemplate attributeTemplate) {
+        return attributeTemplate.getExcelColumnName();
+    }
+}