+++ src/main/java/de/samply/template/graph/factory/ExcelConverterTemplateGraphFactory.java
@@ -1,0 +1,18 @@
+package de.samply.template.graph.factory;
+
+import de.samply.converter.Format;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ExcelConverterTemplateGraphFactory extends ConverterTemplateGraphFactory{
+
+    public ExcelConverterTemplateGraphFactory() {
+        super(new ExcelContainerAttributeKeys());
+    }
+
+    @Override
+    protected Format getOutputFormat() {
+        return Format.EXCEL;
+    }
+
+}