+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2023-10-31]
+## [1.1.0 - 2023-11-07]
 ### Added
 - Java Opts
 - Divide Excel file in several files if number of rows is too large
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch query execution errors
 - Fetch query executions and execution errors by query id and query execution id
 - CSV Separator Replacement
+- Template Graph
 
 ### Changed
 - Move FHIR Packages to dktk-exporter